tests/inputdevice-wheel.html

<!DOCTYPE html>
<title>InputDevice test for wheel-triggered events</title>
<meta name="viewport" content="width=device-width">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>
<style>
#target {
  border: 1px solid black;
  width: 100px;
  height: 100px;
  background-color: lightgreen;
  margin-bottom: 20px;
}
</style>

<h1>InputDevice - test for wheel-triggered events</h1>
<p>Mousewheel over the box below.</p>
<div id="target" tabindex=1></div>
<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceValidator(sourceDevice) {
  assert_true(event.sourceDevice != null, "sourceDevice is set");
  assert_false(event.sourceDevice.firesTouchEvents, "sourcedevice.firesTouchEvents");
}

function runTests() {
  var events = ["wheel"];
  runInputDeviceEventTests(events, inputDeviceValidator);
}
</script>
